Rename `parser.py` to `course_parser.py`
